@@ -1,12 +1,13 @@
 package toyhose
 
 import (
+	"context"
 	"testing"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/service/kinesis"
-	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/credentials"
+	"github.com/aws/aws-sdk-go-v2/service/kinesis"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
 )
 
 var awsConf *aws.Config
@@ -14,33 +15,33 @@ var s3EndpointURL = "http://localhost:9000"
 var kinesisEndpointURL = "http://localhost:4567"
 
 func init() {
-	awsConf = aws.NewConfig().
-		WithRegion("us-east-1").
-		WithCredentials(credentials.NewStaticCredentials("XXXXXXXX", "YYYYYYYY", ""))
+	awsConf = aws.NewConfig()
+	awsConf.Region = "us-east-1"
+	awsConf.Credentials = credentials.NewStaticCredentialsProvider("XXXXXXXX", "YYYYYYYY", "")
 	_ = SetupLogger()
 }
 
-func setupS3(t *testing.T, s3cli *s3.S3, bucket string) error {
-	if _, err := s3cli.CreateBucket(&s3.CreateBucketInput{
+func setupS3(t *testing.T, s3cli *s3.Client, bucket string) error {
+	if _, err := s3cli.CreateBucket(context.Background(), &s3.CreateBucketInput{
 		Bucket: &bucket,
 	}); err != nil {
 		return err
 	}
 	t.Cleanup(func() {
-		_, _ = s3cli.DeleteBucket(&s3.DeleteBucketInput{Bucket: &bucket})
+		_, _ = s3cli.DeleteBucket(context.Background(), &s3.DeleteBucketInput{Bucket: &bucket})
 	})
 	return nil
 }
 
-func setupKinesisStream(t *testing.T, kinCli *kinesis.Kinesis, streamName string, shardCount int64) error {
-	if _, err := kinCli.CreateStream(&kinesis.CreateStreamInput{
+func setupKinesisStream(t *testing.T, kinCli *kinesis.Client, streamName string, shardCount int32) error {
+	if _, err := kinCli.CreateStream(context.Background(), &kinesis.CreateStreamInput{
 		StreamName: &streamName,
 		ShardCount: &shardCount,
 	}); err != nil {
 		return err
 	}
 	t.Cleanup(func() {
-		_, _ = kinCli.DeleteStream(&kinesis.DeleteStreamInput{
+		_, _ = kinCli.DeleteStream(context.Background(), &kinesis.DeleteStreamInput{
 			StreamName: &streamName,
 		})
 	})
