@@ -1,22 +1,23 @@
 package toyhose
 
 import (
+	"context"
+	"crypto/sha256"
 	"errors"
-	"io"
 	"net/http"
 	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	v4 "github.com/aws/aws-sdk-go/aws/signer/v4"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	v4 "github.com/aws/aws-sdk-go-v2/aws/signer/v4"
+	"github.com/aws/smithy-go"
 )
 
 var (
 	errInvalidSignature = errors.New("invalid signature")
 )
 
-func verifyV4(c *aws.Config, req *http.Request, body io.ReadSeeker) error {
+func verifyV4(c *aws.Config, req *http.Request, body []byte) error {
 	refAuth := req.Header.Get("Authorization")
 	sigHeaders := strings.Split(refAuth, ", ")[1]
 	sigHeaderVal := strings.Split(sigHeaders, "=")[1]
@@ -29,11 +30,28 @@ func verifyV4(c *aws.Config, req *http.Request, body io.ReadSeeker) error {
 		}
 	}
 	copiedReq.Header = remainHeaders
-	if _, err := v4.NewSigner(c.Credentials).Sign(copiedReq, body, "firehose", *c.Region, ts); err != nil {
-		return awserr.New("IncompleteSignature", "failed to build sign", err)
+	cred, err := c.Credentials.Retrieve(context.Background())
+	if err != nil {
+		return &smithy.GenericAPIError{
+			Message: "failed to retrieve credentials",
+			Code:    "IncompleteSignature",
+			Fault:   smithy.FaultServer,
+		}
+	}
+	payloadHash := sha256.Sum256(body)
+	if err := v4.NewSigner().SignHTTP(context.Background(), cred, copiedReq, string(payloadHash[:]), "firehose", c.Region, ts); err != nil {
+		return &smithy.GenericAPIError{
+			Message: "failed to build sign",
+			Code:    "IncompleteSignature",
+			Fault:   smithy.FaultClient,
+		}
 	}
 	if a := copiedReq.Header.Get("Authorization"); a != refAuth {
-		return awserr.New("AccessDeniedException", "signature mismatched", errInvalidSignature)
+		return &smithy.GenericAPIError{
+			Message: "signature mismatched",
+			Code:    "IncompleteSignature",
+			Fault:   smithy.FaultClient,
+		}
 	}
 	return nil
 }
